requirements.txt
altair==4.1.0
joblib==1.1.1
matplotlib==3.6.3
numpy==1.23.5
opencv_python==4.7.0.68
pandas==1.5.3
Pillow==9.4.0
pyhdust==1.3.35
pyvista==0.37.0
requests==2.28.2
scikit-image==0.19.3
scipy==1.10.0
seaborn==0.12.2
streamlit==1.18
streamlit_cropper==0.2.1
streamlit_lottie==0.0.3
streamlit_option_menu==0.3.2
tensorflow==2.9.0
ipython
pythreejs
scikit-learn==1.0.2
stpyvista